+++ src/main/java/yuuki1293/pccard/mixins/GenericEntryStackHelperJEIMixin.java
@@ -2,6 +2,7 @@ package yuuki1293.pccard.mixins;
 
 import appeng.api.stacks.GenericStack;
 import appeng.integration.modules.jei.GenericEntryStackHelper;
+import com.gregtechceu.gtceu.common.data.GTItems;
 import mezz.jei.api.gui.ingredient.IRecipeSlotView;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.recipe.RecipeIngredientRole;
@@ -11,7 +12,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import yuuki1293.pccard.CompetitionFixer;
 import yuuki1293.pccard.ConfigClient;
 import yuuki1293.pccard.PCCard;
 
@@ -31,7 +31,7 @@ public abstract class GenericEntryStackHelperJEIMixin {
 
         var inputs = cir.getReturnValue();
 
-        var circuitStack = CompetitionFixer.PC.get().asStack();
+        var circuitStack = GTItems.PROGRAMMED_CIRCUIT.asStack();
         var circuit = recipeLayout.getSlotViews(RecipeIngredientRole.CATALYST).stream()
             .filter(ei -> {
                 var stack = ei.getDisplayedItemStack().orElse(ItemStack.EMPTY);