+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Fixed
  - Does not work properly across multiple pattern providers using crafting co-processing units [#11](https://github.com/yuuki1293/ProgrammedCircuitCard/issues/11)
 
+### Removed
+ - Support for older GregTech versions.
+
 ## [1.0.10] - 2025-02-16
 ### Added
  - JEI integration config. [#8](https://github.com/yuuki1293/ProgrammedCircuitCard/issues/8)